+++ game/dota_addons/barebones/scripts/vscripts/libraries/modifiers/modifier_animation_translate_permanent.lua
@@ -1,4 +1,4 @@
-modifier_animation_translate = class({})
+modifier_animation_translate_permanent = class({})
 
 local _CODE_TO_ANIMATION_TRANSLATE = {
   [0]= "loadout",
@@ -374,7 +374,7 @@ local _CODE_TO_ANIMATION_TRANSLATE = {
   [370]="windwalk",
 }
 
-function modifier_animation_translate:OnCreated(keys) 
+function modifier_animation_translate_permanent:OnCreated(keys) 
   if not IsServer() then
     self.translate = _CODE_TO_ANIMATION_TRANSLATE[keys.stack_count]
   else
@@ -382,23 +382,23 @@ function modifier_animation_translate:OnCreated(keys)
   end
 end
 
-function modifier_animation_translate:GetAttributes()
+function modifier_animation_translate_permanent:GetAttributes()
   return MODIFIER_ATTRIBUTE_PERMANENT + MODIFIER_ATTRIBUTE_IGNORE_INVULNERABLE --+ MODIFIER_ATTRIBUTE_MULTIPLE
 end
 
-function modifier_animation_translate:IsHidden()
+function modifier_animation_translate_permanent:IsHidden()
   return true
 end
 
-function modifier_animation_translate:IsDebuff() 
+function modifier_animation_translate_permanent:IsDebuff() 
   return false
 end
 
-function modifier_animation_translate:IsPurgable() 
+function modifier_animation_translate_permanent:IsPurgable() 
   return false
 end
 
-function modifier_animation_translate:DeclareFunctions() 
+function modifier_animation_translate_permanent:DeclareFunctions() 
   local funcs = {
     MODIFIER_PROPERTY_TRANSLATE_ACTIVITY_MODIFIERS,
   }
@@ -406,13 +406,13 @@ function modifier_animation_translate:DeclareFunctions()
   return funcs
 end
 
-function modifier_animation_translate:GetActivityTranslationModifiers(...)
+function modifier_animation_translate_permanent:GetActivityTranslationModifiers(...)
   --print('TRANSLATE MODIFIER', IsServer())
   --print(self.translate)
   return self.translate or 0
 end
 
---[[function modifier_animation_translate:CheckState() 
+--[[function modifier_animation_translate_permanent:CheckState() 
   local state = {
     [MODIFIER_STATE_UNSELECTABLE] = true,
     [MODIFIER_STATE_INVULNERABLE] = true,